# docs/artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mini-conf API",
    "description": "HTTP-first deployment configuration APIs for mini-conf",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/open/deployment-sync-records": {
      "post": {
        "tags": [
          "open"
        ],
        "operationId": "create_sync_record",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentSyncRecordRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sync record accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentSyncResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Bearer token does not match target deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Deployment, config file, or release not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Database bootstrap disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/open/releases/{revision}": {
      "get": {
        "tags": [
          "open"
        ],
        "operationId": "get_release",
        "parameters": [
          {
            "name": "revision",
            "in": "path",
            "description": "Release revision to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Release content payload",
            "headers": {
              "cache-control": {
                "schema": {
                  "type": "string"
                },
                "description": "Cache policy"
              },
              "etag": {
                "schema": {
                  "type": "string"
                },
                "description": "Entity tag for release content hash"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseContentResponse"
                }
              }
            }
          },
          "304": {
            "description": "Release not modified",
            "headers": {
              "cache-control": {
                "schema": {
                  "type": "string"
                },
                "description": "Cache policy"
              },
              "etag": {
                "schema": {
                  "type": "string"
                },
                "description": "Entity tag for release content hash"
              }
            }
          },
          "400": {
            "description": "Invalid revision path parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Release not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Database bootstrap disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeploymentSyncRecordRequestBody": {
        "type": "object",
        "required": [
          "project",
          "environment",
          "deployment_key",
          "config",
          "action",
          "status",
          "detail"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "config": {
            "type": "string"
          },
          "deployment_key": {
            "type": "string"
          },
          "detail": {
            "type": [
              "object",
              "null"
            ]
          },
          "environment": {
            "type": "string"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "process_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "project": {
            "type": "string"
          },
          "reported_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "revision": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "DeploymentSyncResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ReleaseConfig": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ReleaseContentResponse": {
        "type": "object",
        "required": [
          "release",
          "deployment",
          "config",
          "content",
          "metadata"
        ],
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ReleaseConfig"
          },
          "content": {
            "type": "string"
          },
          "deployment": {
            "$ref": "#/components/schemas/ReleaseDeployment"
          },
          "metadata": {
            "$ref": "#/components/schemas/ReleaseMetadata"
          },
          "release": {
            "$ref": "#/components/schemas/ResolveRelease"
          }
        }
      },
      "ReleaseDeployment": {
        "type": "object",
        "required": [
          "project",
          "environment",
          "deployment_key"
        ],
        "properties": {
          "deployment_key": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "project": {
            "type": "string"
          }
        }
      },
...
      "ReleaseMetadata": {
        "type": "object",
        "properties": {
          "change_summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "schema_version": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ResolveRelease": {
        "type": "object",
        "required": [
          "revision",
          "content_hash",
          "format",
          "published_at",
          "apply_mode"
        ],
        "properties": {
          "apply_mode": {
            "type": "string"
          },
          "content_hash": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "published_at": {
            "type": "string"
          },
          "revision": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      },
      "session_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mini_conf_session"
      }
    }
  },
...
}